+++ sdk/dotnet/GetRepository.cs
@@ -144,6 +144,10 @@ namespace Pulumi.Github
         /// The list of topics of the repository.
         /// </summary>
         public readonly ImmutableArray<string> Topics;
+        /// <summary>
+        /// Whether the repository is public, private or internal.
+        /// </summary>
+        public readonly string Visibility;
 
         [OutputConstructor]
         private GetRepositoryResult(
@@ -189,7 +193,9 @@ namespace Pulumi.Github
 
             string svnUrl,
 
-            ImmutableArray<string> topics)
+            ImmutableArray<string> topics,
+
+            string visibility)
         {
             AllowMergeCommit = allowMergeCommit;
             AllowRebaseMerge = allowRebaseMerge;
@@ -213,6 +219,7 @@ namespace Pulumi.Github
             SshCloneUrl = sshCloneUrl;
             SvnUrl = svnUrl;
             Topics = topics;
+            Visibility = visibility;
         }
     }
 }