+++ sdk/go/github/config/config.go
@@ -8,14 +8,6 @@ import (
 	"github.com/pulumi/pulumi/sdk/v2/go/pulumi/config"
 )
 
-// Authenticate without a token. When `anonymous`is true, the provider will not be able to access resourcesthat require
-// authentication.
-//
-// Deprecated: For versions later than 3.0.0, absence of a token enables this mode
-func GetAnonymous(ctx *pulumi.Context) bool {
-	return config.GetBool(ctx, "github:anonymous")
-}
-
 // The GitHub Base API URL
 func GetBaseUrl(ctx *pulumi.Context) string {
 	v, err := config.Try(ctx, "github:baseUrl")
@@ -25,17 +17,12 @@ func GetBaseUrl(ctx *pulumi.Context) string {
 	return getEnvOrDefault("https://api.github.com/", nil, "GITHUB_BASE_URL").(string)
 }
 
-// Deprecated: For versions later than 3.0.0, absence of an organization enables this mode
-func GetIndividual(ctx *pulumi.Context) bool {
-	return config.GetBool(ctx, "github:individual")
-}
-
-// Whether server should be accessed without verifying the TLS certificate.
+// Enable `insecure` mode for testing purposes
 func GetInsecure(ctx *pulumi.Context) bool {
 	return config.GetBool(ctx, "github:insecure")
 }
 
-// The GitHub organization name to manage. If `individual` is false, `organization` is required.
+// The GitHub organization name to manage. Use this field instead of `owner` when managing organization accounts.
 func GetOrganization(ctx *pulumi.Context) string {
 	v, err := config.Try(ctx, "github:organization")
 	if err == nil {
@@ -44,7 +31,12 @@ func GetOrganization(ctx *pulumi.Context) string {
 	return getEnvOrDefault("", nil, "GITHUB_ORGANIZATION").(string)
 }
 
-// The OAuth token used to connect to GitHub. If `anonymous` is false, `token` is required.
+// The GitHub owner name to manage. Use this field instead of `organization` when managing individual accounts.
+func GetOwner(ctx *pulumi.Context) string {
+	return config.Get(ctx, "github:owner")
+}
+
+// The OAuth token used to connect to GitHub. `anonymous` mode is enabled if `token` is not configured.
 func GetToken(ctx *pulumi.Context) string {
 	v, err := config.Try(ctx, "github:token")
 	if err == nil {