+++ sdk/go/github/repositoryWebhook.go
@@ -11,10 +11,7 @@ import (
 )
 
 // This resource allows you to create and manage webhooks for repositories within your
-// GitHub organization.
-//
-// This resource cannot currently be used to manage webhooks for *personal* repositories,
-// outside of organizations.
+// GitHub organization or personal account.
 //
 // ## Example Usage
 //