+++ sdk/nodejs/getRepository.ts
@@ -132,4 +132,8 @@ export interface GetRepositoryResult {
      * The list of topics of the repository.
      */
     readonly topics: string[];
+    /**
+     * Whether the repository is public, private or internal.
+     */
+    readonly visibility: string;
 }