+++ sdk/python/pulumi_github/get_repository.py
@@ -19,7 +19,7 @@ class GetRepositoryResult:
     """
     A collection of values returned by getRepository.
     """
-    def __init__(__self__, allow_merge_commit=None, allow_rebase_merge=None, allow_squash_merge=None, archived=None, default_branch=None, description=None, full_name=None, git_clone_url=None, has_downloads=None, has_issues=None, has_projects=None, has_wiki=None, homepage_url=None, html_url=None, http_clone_url=None, id=None, name=None, node_id=None, private=None, ssh_clone_url=None, svn_url=None, topics=None):
+    def __init__(__self__, allow_merge_commit=None, allow_rebase_merge=None, allow_squash_merge=None, archived=None, default_branch=None, description=None, full_name=None, git_clone_url=None, has_downloads=None, has_issues=None, has_projects=None, has_wiki=None, homepage_url=None, html_url=None, http_clone_url=None, id=None, name=None, node_id=None, private=None, ssh_clone_url=None, svn_url=None, topics=None, visibility=None):
         if allow_merge_commit and not isinstance(allow_merge_commit, bool):
             raise TypeError("Expected argument 'allow_merge_commit' to be a bool")
         pulumi.set(__self__, "allow_merge_commit", allow_merge_commit)
@@ -86,6 +86,9 @@ class GetRepositoryResult:
         if topics and not isinstance(topics, list):
             raise TypeError("Expected argument 'topics' to be a list")
         pulumi.set(__self__, "topics", topics)
+        if visibility and not isinstance(visibility, str):
+            raise TypeError("Expected argument 'visibility' to be a str")
+        pulumi.set(__self__, "visibility", visibility)
 
     @property
     @pulumi.getter(name="allowMergeCommit")
@@ -254,6 +257,14 @@ class GetRepositoryResult:
         """
         return pulumi.get(self, "topics")
 
+    @property
+    @pulumi.getter
+    def visibility(self) -> str:
+        """
+        Whether the repository is public, private or internal.
+        """
+        return pulumi.get(self, "visibility")
+
 
 class AwaitableGetRepositoryResult(GetRepositoryResult):
     # pylint: disable=using-constant-test
@@ -282,7 +293,8 @@ class AwaitableGetRepositoryResult(GetRepositoryResult):
             private=self.private,
             ssh_clone_url=self.ssh_clone_url,
             svn_url=self.svn_url,
-            topics=self.topics)
+            topics=self.topics,
+            visibility=self.visibility)
 
 
 def get_repository(full_name: Optional[str] = None,
@@ -335,4 +347,5 @@ def get_repository(full_name: Optional[str] = None,
         private=__ret__.private,
         ssh_clone_url=__ret__.ssh_clone_url,
         svn_url=__ret__.svn_url,
-        topics=__ret__.topics)
+        topics=__ret__.topics,
+        visibility=__ret__.visibility)